test(proxy): move compact trigger coverage into focused module
